Adding GitHub OAuth (discontinue)
We cannot go foward because of CORS. That is, oidc-client[-ts] assumes
that the IdP configured CORS correctly, but GitHub's OAuth endpoint does
not support CORS. So, we should implement it in server side. Although
implementing in server side requires network access to run the app,
GitHub access is also required in the first place.